crud de modulos y permisos

+++ src/main/java/com/alexander/sistema_cerro_verde_backend/repository/UsuariosRepository.java
@@ -1,9 +1,10 @@
 package com.alexander.sistema_cerro_verde_backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import com.alexander.sistema_cerro_verde_backend.entity.Usuarios;
 
 @Repository
-public interface UsuariosRepository extends JpaRepository<Usuarios, Long> {
+public interface UsuariosRepository extends JpaRepository<Usuarios, Integer> {
     public Usuarios findByUsername(String username);
 }
